Extract formula engine test utils into a separate package
